+++ blocks/cards/cards.js
@@ -1,4 +1,4 @@
-import { createOptimizedPicture } from '../../scripts/aem.js';
+import { createOptimizedPicture, waitForFirstImage } from '../../scripts/aem.js';
 
 export default async function decorate(block) {
   /* change to ul, li */
@@ -12,7 +12,8 @@ export default async function decorate(block) {
     });
     ul.append(li);
   });
-  ul.querySelectorAll('picture > img').forEach((img) => img.closest('picture').replaceWith(createOptimizedPicture(img.src, img.alt, false, [{ width: '750' }])));
+  ul.querySelectorAll('picture > img').forEach((img) => img.closest('picture').replaceWith(createOptimizedPicture(img.src, img.alt, true, [{ width: '750' }])));
   block.textContent = '';
   block.append(ul);
+  await waitForFirstImage(block);
 }
